refactor: 💡 flatten tokens folders, add avatar provider
