+++ design-toolkit/components/src/components/avatar/index.tsx
@@ -12,7 +12,7 @@
 'use client';
 
 import 'client-only';
-import { spacingXs } from '@/tokens/generated/tokens';
+import { spacingXs } from '@/tokens/';
 import { Person } from '@accelint/icons';
 import { Fallback, Image, Root } from '@radix-ui/react-avatar';
 import { createContext } from 'react';
@@ -20,12 +20,19 @@ import { type ContextValue, useContextProps } from 'react-aria-components';
 import { Badge } from '../badge';
 import { Icon } from '../icon';
 import { AvatarStyles, AvatarStylesDefaults } from './styles';
-import type { AvatarProps } from './types';
+import type { AvatarProps, AvatarProviderProps } from './types';
 
 const { avatar, image, fallback, content } = AvatarStyles();
 
 export const AvatarContext =
   createContext<ContextValue<AvatarProps, HTMLSpanElement>>(null);
+
+function AvatarProvider({ children, ...props }: AvatarProviderProps) {
+  return (
+    <AvatarContext.Provider value={props}>{children}</AvatarContext.Provider>
+  );
+}
+AvatarProvider.displayName = 'Avatar.Provider';
 
 export function Avatar({ ref, ...props }: AvatarProps) {
   [props, ref] = useContextProps(props, ref ?? null, AvatarContext);
@@ -70,3 +77,4 @@ export function Avatar({ ref, ...props }: AvatarProps) {
   );
 }
 Avatar.displayName = 'Avatar';
+Avatar.Provider = AvatarProvider;
