+++ design-toolkit/components/src/index.ts
@@ -29,7 +29,10 @@ export type {
 } from './components/accordion/types';
 export { Avatar, AvatarContext } from './components/avatar';
 export { AvatarStyles, AvatarStylesDefaults } from './components/avatar/styles';
-export type { AvatarProps } from './components/avatar/types';
+export type {
+  AvatarProps,
+  AvatarProviderProps,
+} from './components/avatar/types';
 export { Badge, BadgeContext } from './components/badge';
 export { BadgeStyles, BadgeStylesDefaults } from './components/badge/styles';
 export type { BadgeProps, BadgeProviderProps } from './components/badge/types';
@@ -333,5 +336,5 @@ export {
   typographyHeaderXsSize,
   typographyHeaderXxlHeight,
   typographyHeaderXxlSize,
-} from './tokens/generated/tokens';
-export type { RGBAColor } from './tokens/generated/tokens';
+} from './tokens';
+export type { RGBAColor } from './tokens';
